rename pages to components
