+++ src/components/DailyIncreaseChart/dailyIncrease.js
