+++ src/components/DailyIncreaseChart/index.js
