+++ src/components/Header/drawPageTitleCount.js
