+++ src/components/Kpi/drawKpis.js
