+++ src/components/SpreadTrendChart/spreadTrend.js
